Add some logging.
Currently getting a crash with the file I'm using where the number of hits in the nMCStepPositions variable is two orders of magnitude too big.
Hard-coding that to check the rest of the code produces an issue when attempting to fill the nMCStepMomentas variable, such that the second loop is never ran.