Remove ConsolidatedHitsBaseTool.
For now, this doesn't need or make sense to be a tool.
It can be added in the same way as the iterative method.
In the future it may make sense to split this, future iterations and
the iterative treatment of into a different class of "clean up tools",
that are used to take the created hits and augment and tidy them.