Refactor cluster dumping to allow running over view names, not cluster.
Cluster lists have been split in two, tracks and showers.
This way, you just specify the views you care about and get both.

May be worth extending to allow the origin list to be known as well,
which would allow more targeted plots as well, which focus on the things
which can actually be improved (i.e. ignore the track list).