Add additional / updated error checking.
The inclusion of hit filtering changes some parts of the code, with
previously invalid states now being potentially valid for a single view,
when filtering is applied. Rather than throwing errors in those cases,
alert the user when appropriate, but continue and try the other views.
We don't want to have the DL Vertexing throw errors for valid cases such
as failing to find a vertex in a fully filtered cosmic ray slice etc.